backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
anthropic
transformers
torch
scikit-learn
pandas
numpy
python-dotenv